requirements.txt
# Cell Cycle Prediction - Python Dependencies
# ============================================
# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0
imbalanced-learn>=0.9.0

# Deep Learning - PyTorch
torch>=1.10.0
torchmetrics>=0.7.0

# Hyperparameter Optimization
optuna>=3.0.0

# Model Interpretability
shap>=0.40.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Data Processing
joblib>=1.1.0

# Additional utilities
tqdm>=4.62.0
pyyaml>=6.0

# For notebook conversion and development
jupyter>=1.0.0
ipykernel>=6.0.0
